requirements.txt
Django==4.2.13
django-appconf==1.0.5
django-auditlog==3.0.0
django-cors-headers==3.14.0
django-countries==7.2.1
django-environ==0.4.5
django-extensions==3.2.1
django-fsm==2.8.1
django-fsm-log==3.1.0
django-rest-framework==0.1.0
django-storages==1.12.3
django-subscriptions==2.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==5.4.1
razorpay==1.3.0
redis==4.5.3
requests==2.26.0
whitenoise==4.1.4
yarl==1.8.2
zcrmsdk==3.1.0
